Fixing rendering issue in Galaxy S2.
In OpenGLES, indices need to be in uint16_t/uint8_t.
The fix switches all indices to uint16_t.  In other devices,
it's working even it's not under GLES spec but Mali driver
in the device gracefully catches the issue and is failing at
glDrawElements().

Tested: On N5, Galaxy S2
Bug: 18071449

+++ src/splat_game.cpp
@@ -41,7 +41,7 @@ namespace splat {
 static const int kQuadNumVertices = 4;
 static const int kQuadNumIndices = 6;
 
-static const int kQuadIndices[] = { 0, 1, 2, 2, 1, 3 };
+static const unsigned short kQuadIndices[] = { 0, 1, 2, 2, 1, 3 };
 
 static const Attribute kQuadMeshFormat[] =
     { kPosition3f, kTexCoord2f, kNormal3f, kTangent4f, kEND };
